requirements.txt
langchain==0.3.19
langchain-groq==0.2.4
langchain-community==0.3.18
wikipedia==1.4.0
langgraph==0.3.0
sentence-transformers==3.4.1
faiss-gpu==1.7.2
youtube-search==2.1.2
youtube-search-python==1.6.6
toml==0.10.2
loguru==0.7.3
fastapi==0.115.12
uvicorn==0.34.0
jinja2==3.0.2
passlib==1.7.4
python-jose==3.4.0
python-multipart==0.0.20
asyncpg==0.30.0